Quasi Information Criterion

Function for calculating the quasi-likelihood under the independence model information criterion (QIC), quasi-likelihood, correlation information criterion (CIC), and corrected QIC for one or several fitted geeglm model object from the geepack package.

# S3 method for geeglm
QIC(object, tol = .Machine$double.eps, ...)

# S3 method for ordgee
QIC(object, tol = .Machine$double.eps, ...)

# S3 method for geekin
QIC(object, tol = .Machine$double.eps, ...)

QIC(object, tol = .Machine$double.eps, ...)

Arguments

object
a fitted GEE model from the geepack package. Currently only works on geeglm objects

tol
the tolerance used for matrix inversion

…
optionally more fitted geeglm model objects

Value

A vector or matrix with the QIC, QICu, quasi likelihood, CIC, the number of mean effect parameters, and the corrected QIC for each GEE object

Details

QIC is used to select a correlation structure. The QICu is used to compare models that have the same working correlation matrix and the same quasi-likelihood form but different mean specifications. CIC has been suggested as a more robust alternative to QIC when the model for the mean may not fit the data very well and when models with different correlation structures are compared.

Models with smaller values of QIC, CIC, QICu, or QICC are preferred.

If the MASS package is loaded then the ginv function is used for matrix inversion. Otherwise the standard solve function is used.

References

Pan, W. (2001). Akaike's information criterion in generalized estimating equations. Biometrics, 57, 120-125. Hardin, J.W. and Hilbe, J.M. (2012). Generalized Estimating Equations, 2nd Edition, Chapman and Hall/CRC: New York. Hin, L.-Y. and Wang, Y-G. (2009). Working-correlation-structure identification in generalized estimating equations, Statistics in Medicine 28: 642-658. Thall, P.F. and Vail, S.C. (1990). Some Covariance Models for Longitudinal Count Data with Overdispersion. Biometrics, 46, 657-671.

See also

geeglm

Examples

library(geepack)
data(ohio)
fit <- geeglm(resp ~ age + smoke + age:smoke, id=id, data=ohio,
             family=binomial, corstr="exch", scale.fix=TRUE)
QIC(fit)

#>         QIC        QICu   Quasi Lik         CIC      params        QICC 
#> 1827.489654 1827.480050 -909.740025    4.004802    4.000000 1827.517665