Bee data. Number of different types of bees caught.

Number of different types of bees caught in plates of different colours. There are four locations and within each location there are three replicates consisting of three plates of the three different colours (yellow, white and blue). Data are collected at 5 different dates over the summer season. Only data from one date available until data has been published.

Format

A data frame with 72 observations on the following 7 variables.

Locality

a factor with levels Havreholm Kragevig Saltrup Svaerdborg. Four different localities in Denmark.

Replicate

a factor with levels A B C

Color

a factor with levels Blue White Yellow. Colour of plates

Time

a factor with levels july1 july14 june17 june3 june6. Data collected at different dates in summer season. Only one day is present in the current data frame until the full data has been released.

Type

a factor with levels Bumblebees Solitary. Type of bee.

Number

a numeric vector. The response variable with number of bees catched.

id

a numeric vector. The id of the clusters (each containg three plates).

Source

Data were kindly provided by Casper Ingerslev Henriksen, Department of Agricultural Sciences, KU-LIFE. Added by Torben Martinussen <tma@life.ku.dk>

Examples

data(bees)
model <- glm(Number ~ Locality + Type*Color,
             family=poisson, data=bees)