Blood clotting for 158 rats

Blood clotting activity (PCA) is measured for 158 Norway rats from two locations just before (baseline) and four days after injection of an anticoagulant (bromadiolone). Normally this would cause reduced blood clotting after 4 days compared to the baseline, but these rats are known to possess anticoagulent resistence to varying extent. The purpose is to relate anticoagulent resistence to gender and location and perhaps weight. Dose of injection is, however, admistered according to weight and gender.

Format

A data frame with 158 observations on the following 6 variables.

rat

a numeric vector

locality

a factor with levels Loc1 Loc2

sex

a factor with levels F M

weight

a numeric vector

PCA0

a numeric vector with percent blood clotting activity at baseline

PCA4

a numeric vector with percent blood clotting activity on day 4

Source

Ann-Charlotte Heiberg, project at The Royal Veterinary and Agricultural University, 1999. Added by Ib M. Skovgaard <ims@life.ku.dk>

Examples

 data(clotting)
 dim(clotting)

#> [1] 158   6

 head(clotting)

#>   rat locality sex weight PCA0 PCA4
#> 1   1     Loc1   F    284 78.6 73.2
#> 2   2     Loc1   F    274 65.2 67.8
#> 3   3     Loc1   F    276 78.2 91.9
#> 4   4     Loc1   F    298 62.9 76.3
#> 5   5     Loc1   F    284 55.3 53.1
#> 6   6     Loc1   F    266 77.7 80.1

 day0= transform(clotting, day=0, pca=PCA0)
 day4= transform(clotting, day=4, pca=PCA4)
 day.both= rbind(day0,day4)
 m1= lm(pca ~ rat + day*locality + day*sex, data=day.both)
 anova(m1)

#> Analysis of Variance Table
#> 
#> Response: pca
#>               Df Sum Sq Mean Sq F value    Pr(>F)    
#> rat            1   3090    3090  1.5599   0.21263    
#> day            1  12083   12083  6.0996   0.01406 *  
#> locality       1  86364   86364 43.5985 1.757e-10 ***
#> sex            1    679     679  0.3430   0.55852    
#> day:locality   1    309     309  0.1561   0.69301    
#> day:sex        1   5339    5339  2.6953   0.10166    
#> Residuals    309 612094    1981                      
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

 summary(m1)

#> 
#> Call:
#> lm(formula = pca ~ rat + day * locality + day * sex, data = day.both)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -102.075  -25.065   -5.473   12.944  307.721 
#> 
#> Coefficients:
#>                  Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)       78.2248     7.5826  10.316  < 2e-16 ***
#> rat                0.4522     0.2120   2.133  0.03371 *  
#> day               -0.8559     2.0795  -0.412  0.68093    
#> localityLoc2     -55.9472    18.0586  -3.098  0.00213 ** 
#> sexM              14.0545    11.0268   1.275  0.20342    
#> day:localityLoc2  -0.9210     2.5049  -0.368  0.71337    
#> day:sexM          -4.1593     2.5335  -1.642  0.10166    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 44.51 on 309 degrees of freedom
#> Multiple R-squared:  0.1498,	Adjusted R-squared:  0.1333 
#> F-statistic: 9.075 on 6 and 309 DF,  p-value: 3.78e-09
#> 

 m2= lm(pca ~ rat + day, data=day.both)
 anova(m2)

#> Analysis of Variance Table
#> 
#> Response: pca
#>            Df Sum Sq Mean Sq F value  Pr(>F)  
#> rat         1   3090  3089.9  1.3723 0.24232  
#> day         1  12083 12082.6  5.3660 0.02118 *
#> Residuals 313 704786  2251.7                  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

## Log transformation suggested.
## Random effect of rat.
## maybe str(clotting) ; plot(clotting) ...