Drop All Possible Single Terms to a geeglm Model Using Wald or Score Test

Compute all the single terms in the scope argument that can dropped from the model, and compute a table of the corresponding Wald test statistics.

# S3 method for geeglm
drop1(object, scope, test = c("Wald", "none", "score",
  "sasscore"), method = c("robust", "naive", "sandwich"), ...)

Arguments

object
a fitted object of class geese.

scope
a formula giving the terms to be considered for adding or dropping.

test
the type of test to include.

method
Indicates which method is used for computing the standard error. robust is the default and corresponds to the modified sandwich estimator. naive is the classical naive cariance estimate. sandwich is an alias for robust.

…
other arguments. Not currently used

Value

An object of class "anova" summarizing the differences in fit between the models.

See also

drop1, geeglm, geese

Examples

library(geepack)
data(ohio)
fit <- geeglm(resp ~ age + smoke + age:smoke, id=id, data=ohio,
             family=binomial, corstr="exch", scale.fix=TRUE)
drop1(fit)

#> Single term deletions
#> 
#> Model:
#> resp ~ age + smoke + age:smoke
#>           DF    Wald Pr(>Chi)
#> age:smoke  1 0.64379   0.4223