Gene expression data from two-color dye-swap experiment

Gene expression levels from two-color dye-swap experiment on 6 microarrays. Arrays 1 and 2 represent the first biological sample (ie, the first dye swap), 3 and 4 the second, and arrays 5 and 6 the third.

Format

A data frame with 258000 observations on the following 5 variables.

color

a factor with levels green red representing the dye used for the gene expression

array

a factor with levels 1 2 3 4 5 6 corresponding to the 6 arrays

gene

a factor with 21500 levels representing the genes on the arrays

plant

a factor with levels rnt wt for the two types of plants: runts and wild type

signal

a numeric vector with the gene expression level (normalized but not log transformed)

Source

Data provided by Soren Bak <bak@life.ku.dk>. Added by Claus Ekstrom <ekstrom@sund.ku.dk>

References

Morant, M. et al. (2010). Metabolomic, Transcriptional, Hormonal and Signaling Cross-Talk in Superroot2. Molecular Plant. 3, p.192--211.

Examples

data(superroot2)
# Select one gene
g1 <- superroot2[superroot2$gene=="AT2G24000.1",]
model <- lm(log(signal) ~ plant + color + array, data=g1)
summary(model)

#> 
#> Call:
#> lm(formula = log(signal) ~ plant + color + array, data = g1)
#> 
#> Residuals:
#>        1        2        3        4        5        6   129001   129002 
#>  0.11605  0.11167 -0.13984 -0.07720  0.02379 -0.03447 -0.11605 -0.11167 
#>   129003   129004   129005   129006 
#>  0.13984  0.07720 -0.02379  0.03447 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  6.39677    0.13318  48.032 1.12e-06 ***
#> plantwt     -0.04857    0.09417  -0.516 0.633236    
#> colorred     0.17805    0.09417   1.891 0.131649    
#> array2      -1.55889    0.16311  -9.557 0.000669 ***
#> array3      -0.71928    0.16311  -4.410 0.011603 *  
#> array4      -0.38153    0.16311  -2.339 0.079453 .  
#> array5      -0.80767    0.16311  -4.952 0.007752 ** 
#> array6      -0.46209    0.16311  -2.833 0.047202 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.1631 on 4 degrees of freedom
#> Multiple R-squared:  0.9643,	Adjusted R-squared:  0.9018 
#> F-statistic: 15.43 on 7 and 4 DF,  p-value: 0.00946
#>